% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chpt}
\alias{chpt}
\title{Fits multiple step models for a given trace}
\usage{
chpt(y, times, w)
}
\arguments{
\item{y}{RNA unwinding trace}

\item{times}{Time points at which RNA trace was recorded}

\item{w}{Sequence of window sizes}

\item{cutoff}{Upper percentile for choosing plausible step locations. Default is 0.9 i.e. only top 10\% highest values are choosen by default}

\item{cor}{Default is TRUE indicating that the noise is correlated. If FALSE the noise is assumed uncorrelated}
}
\value{
Returns a list with the following elements
\itemize{
\item results - is a list of length equal to number of models fitted. Each element in this list comprises of elements returned in get.model
\item chpt0 - A list with y, times, w and zstat - the statistics based on which the model is fitted
}
}
\description{
Fits multiple step models for a given trace
}
\examples{
\dontrun{
w<-c(seq(10,90,by=10),seq(100,1000,by=25))
times<-RNA[,2]
chpt1<-chpt(y,times,w)
plot.step(chpt1)
# to save as a pdffile
# plot.step(chpt1,pdfname="RNAFIG1.pdf")
summary(chpt1) # to get bic fit
summary(chpt1,type="aic") # to get aic fit
summary(chpt1,type=5) # to get fit summary of model number 5
plot(chpt1)
mymodel<-get.model(chpt1)
names(mymodel)
get.location(chpt1)
}
}
\references{
Arunajadai SG, Cheng W (2013) Step Detection in Single-Molecule Real Time Trajectories Embedded in Correlated Noise. PLoS ONE 8(3): e59279.

Cheng W, Arunajadai SG, Moffitt JR, Tinoco I Jr, Bustamante C. Single-base pair unwinding and asynchronous RNA release by the hepatitis C virus NS3 helicase. Science. 2011 Sep 23;333(6050):1746-9. doi: 10.1126/science.1206023.
}
\seealso{
get.model, plot.chpt, summary.chpt, get.location
}

