% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.location}
\alias{get.location}
\title{Get step locations}
\usage{
get.location(object, type = c("bic", "aic"))
}
\arguments{
\item{object}{Object of class chpt returned by the chpt() command}

\item{type}{Either "bic" (default), "aic" or and integer between 1 and number of models fit.}
}
\value{
Returns the step locations of the selected model
}
\description{
Get step locations
}
\examples{
\dontrun{
# See example in chpt()
}
}

