% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.model}
\alias{get.model}
\title{Get model characteristics}
\usage{
get.model(object, type = c("bic", "aic"))
}
\arguments{
\item{object}{Object of class chpt returned by the chpt() command}

\item{type}{Either "bic" (default), "aic" or and integer between 1 and number models fit.}
}
\value{
Returns a list with the following elements
\itemize{
\item betas - Estimate of the step size
\item betas.vcov - Estimate of the variance covariance matrix of the step size estimates
\item tstat - The t-statistics for the step estimates
\item fit - the fitted  RNA trace
\item res - the residuals
\item sigma2 - Estimated variance of the underlying noise
\item pval - the P-value for the steps
\item aic - The AIC statistic for the fit
\item bic - The BIC statistic for the fit
\item AR-order - Order of the auto-regression model fitted to the noise
\item AR-coefs - The coefficients from the autoregression model fitted to the noise
}
}
\description{
Get model characteristics
}
\examples{
\dontrun{
# See example in chpt()
}
}
\seealso{
plot.chpt, summary.chpt, get.location
}

