% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.chpt}
\alias{plot.chpt}
\title{Plot the fitted step}
\usage{
plot.chpt(object, type = c("bic", "aic"), pdfname = NULL)
}
\arguments{
\item{object}{Object of class chpt returned by the chpt() command}

\item{type}{Either "bic" (default), "aic" or and integer between 1 and number models fit.}

\item{pdfname}{Name of pdf file with .pdf extension, Default is NULL i.e. no pdf is created}
}
\value{
Plots the steps from the specified model
}
\description{
Plot the fitted step
}
\examples{
\dontrun{
# See example in chpt()
}
}
\seealso{
summary.chpt, get.location
}

