% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RNA}
\alias{RNA}
\title{Simulated RNA trace}
\format{A data frame with 2932 rows and 3 variables

The variables are as follows:
\itemize{
  \item EXT Optical tweezer extension data
  \item Time Time points of measurement
  \item Force Force measured
}}
\description{
A dataset containing simulated RNA unwinding trace
for 2932 time points.
}

