\name{rnasteps-package}
\alias{rnasteps-package}
\alias{rnasteps}
\docType{package}
\title{
Fit Steps for RNA unwiding traces
}
\description{
Single-molecule real time trajectories are embedded in high noise. To extract kinetic or dynamic information of the molecules from these trajectories often requires idealization of the data in steps and dwells. Generalized Least Squares (GLS) is used to detect steps and dwells
in high frequency single-molecule unwinding traces.
}
\details{
\tabular{ll}{
Package: \tab rnasteps\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-12-01\cr
License: \tab GPL-2\cr
}
The main command is chpt() which fits step function to the RNA unwinding trace.
}
\author{
Srikesh Arunajadai

Maintainer: Srikesh Arunajadai<sa2658@columbia.edu>
}
\references{
Arunajadai SG, Cheng W (2013) Step Detection in Single-Molecule Real Time Trajectories Embedded in Correlated Noise. PLoS ONE 8(3): e59279. doi:10.1371/journal.pone.0059279
}
