% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.chpt}
\alias{summary.chpt}
\title{Summary of the fitted steps from chpt command}
\usage{
\method{summary}{chpt}(object, type = c("bic", "aic"))
}
\arguments{
\item{object}{Object of class chpt returned by the chpt() command}

\item{type}{Either "bic" (default), "aic" or and integer between 1 and number models fit.}
}
\value{
coefficients  - A matrix with Step size estimate, standard error, t-statistics and p-value
}
\description{
Gives the summary of the fitted steps based on crieria or model number
}
\examples{
\dontrun{
# See example in chpt()
}
}
\seealso{
plot.chpt, get.location
}

